chore: 패키지명 수정 및 리패키징 global/exception -> global/error
