+++ src/main/java/com/example/short_kki/domain/auth/service/AuthService.java
@@ -7,8 +7,8 @@ import com.example.short_kki.domain.member.entity.OAuthProvider;
 import com.example.short_kki.domain.member.repository.MemberRepository;
 import com.example.short_kki.global.auth.jwt.JwtTokenProvider;
 import com.example.short_kki.global.config.OAuth2Properties;
-import com.example.short_kki.global.exception.BusinessException;
-import com.example.short_kki.global.exception.ErrorCode;
+import com.example.short_kki.global.error.exception.BusinessException;
+import com.example.short_kki.global.error.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
