+++ src/main/java/com/example/short_kki/domain/feed/service/FeedService.java
@@ -9,8 +9,8 @@ import com.example.short_kki.domain.group.repository.GroupRepository;
 import com.example.short_kki.domain.group.repository.GroupMemberRepository;
 import com.example.short_kki.domain.member.entity.Member;
 import com.example.short_kki.domain.member.repository.MemberRepository;
-import com.example.short_kki.global.exception.ErrorCode;
-import com.example.short_kki.global.exception.NotFoundException;
+import com.example.short_kki.global.error.ErrorCode;
+import com.example.short_kki.global.error.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
