+++ src/main/java/com/example/short_kki/domain/recipe/service/RecipeService.java
@@ -14,8 +14,8 @@ import com.example.short_kki.domain.recipe.entity.RecipeStep;
 import com.example.short_kki.domain.recipe.repository.RecipeIngredientRepository;
 import com.example.short_kki.domain.recipe.repository.RecipeRepository;
 import com.example.short_kki.domain.recipe.repository.RecipeStepRepository;
-import com.example.short_kki.global.exception.BusinessException;
-import com.example.short_kki.global.exception.ErrorCode;
+import com.example.short_kki.global.error.exception.BusinessException;
+import com.example.short_kki.global.error.ErrorCode;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
