+++ src/main/java/com/example/short_kki/global/error/exception/BadRequestException.java
@@ -1,0 +1,14 @@
+package com.example.short_kki.global.error.exception;
+
+import com.example.short_kki.global.error.ErrorCode;
+
+public class BadRequestException extends BusinessException {
+
+    public BadRequestException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public BadRequestException(String message) {
+        super(ErrorCode.INVALID_INPUT_VALUE, message);
+    }
+}
