--- src/main/java/com/example/short_kki/global/exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package com.example.short_kki.global.exception;
-
-public class BadRequestException extends BusinessException {
-
-    public BadRequestException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
