--- src/main/java/com/example/short_kki/global/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package com.example.short_kki.global.exception;
-
-public class NotFoundException extends BusinessException {
-
-    public NotFoundException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
